Checkpoint FR-10 netcode/snapshot/event and Vulkan parity work
Snapshot the in-flight progressive networking work (canonical snapshot
timeline, typed event runtime, native transport shadow adapters, rewind
collision provider) together with the Vulkan renderer parity slices.

Build verified clean: meson compile -C builddir-win (107/107, exit 0).

+++ inc/common/bsp.h
@@ -121,7 +121,10 @@ typedef struct mface_s {
     uint16_t        lm_height;
 
     int             drawflags; // DSURF_PLANEBACK, etc
-    int             statebits;
+    // Renderer state includes GLS_REFRACT_ENABLE and later shader controls
+    // above bit 31, so this must retain the full state mask between BSP
+    // registration and the eventual world draw.
+    uint64_t        statebits;
     int             firstvert;
     int             firstbasis;
     uint16_t        light_s, light_t;
